Better CKEditor Widget integration
This does away with the uneeded <mediawrapper> and lets CKEditor work
directly on the media element. This means we do not have to worry about
where the media item is being placed as CKEditor knows how to deal with
the HTML element correctly where as with <mediawrapper> we could tell if
the media was valid where it was inserted.

This also protects unrendered media tokens as there are times when a
subsitute cannot be provided but that does not mean the tag is not
valid.