Adjustments dialog - Allow setting size and CSS class
Because this is a widget the element based CKEditor styles do not
function with it. Ones specific to the widget 'mediabox' can be defined
but these would work for any embed file entity and display mode which is
not ideal. E.g. because classes intended for the image tag are likley
not suited to a file link. The new dialog works around this limitation
by gathering suitable styles and applying them using widget.applyStyle()
which does not validate. This does not defeat CKEditor's ACF as we only
present styles that are defined as allowed anyway.