CKEditor plugin: Use our own method to convert the tokens to markup
Turns out that if the token contains the "link_text" property it will
cause media.filter.create_element() to replace content within any <a>
tag with the link text. This meant if the display contained any links,
the link content in the preview got replaced with the link text (e.g.
images disapeard). The only way to workaround this without making a lot
of changes to media_wysiwyg.filter.js for this editor's use case was to
do our own token conversion. I was wanting to do that anyway ;)